cc: give a character constant the value its encoding prefix implies
CI went red on aarch64 (both the Linux and macOS runners) on two tests, for
two different reasons.

`codegen_char_literal_signedness` was a real defect, and mine. C17
6.4.4.4p10 gives an unprefixed character constant the value of a `char`
object holding the character, converted to `int` -- so it follows plain
`char`'s signedness. Making that signedness target-correct left the
constant behind: on aarch64 c17 said `'\x80'` is -128 while `(char)'\x80'`
is 128, so `enum { PROTO = '\x80' }` no longer matched a `char` read from a
buffer. That is exactly the CPython pickle shape the test was written for,
and it is the sort of thing that links and runs and silently takes the
wrong switch arm. gcc says 128 on aarch64 and -128 on x86-64; c17 now
agrees with both.

The same conversion was wrong for every *prefixed* constant, independently
of any of this and on every target: five consumers each re-derived the
value from a `char` with a hardcoded `as u8 as i8`, so `L'\x80'` was -128
where gcc says 128, `u'\x80'` was 65408, and `L'\xe9'` was -23 where gcc
says 233. The conversion now happens once, in the parser, where the
encoding prefix is still known, and `ExprKind::CharLit` carries the value
rather than a `char` for others to reinterpret. A universal character name
still keeps its code point -- `'é'` is 233 and `'\U0001F600'` is
128512 -- which is c17's deliberate choice and which truncating to a byte
would have destroyed.

`codegen_ptr_arith_narrow_index` was a test that assumed a target: it used
plain `char` to hold -2 as a negative array index. Where `char` is unsigned
that is 254 and the subscript leaves the array, and cross-gcc fails the
same program on aarch64. The negative sections say `signed char` now, which
is what "negative char index" means; an unsigned-index section is added so
the two directions stay pinned.

Verified by executing both programs on both targets -- x86-64 natively and
aarch64 under qemu via the cross toolchain -- against gcc, all four rows
agreeing. Then swept every C program in the test suite the same way: 243
built for aarch64, 241 agree with gcc exactly, and no test now depends on
the host's char signedness. The two that remain are pre-existing and
recorded, not regressions: aarch64 zero-width bit-field layout, and a test
that counts c17's own `__has_builtin` inventory.

+++ cc/ir/linearize.rs
@@ -4902,7 +4902,7 @@ impl<'a> Linearizer<'a> {
 
             ExprKind::CharLit(c) => {
                 let typ = self.expr_type(expr);
-                self.emit_const(*c as u8 as i8 as i128, typ)
+                self.emit_const(*c as i128, typ)
             }
 
             ExprKind::StringLit(s) => {
