+++ cc/parse/parser.rs
@@ -5204,7 +5204,7 @@ impl Parser<'_> {
         match &expr.kind {
             ExprKind::IntLit(val) => Some(*val as i128),
             ExprKind::Int128Lit(val) => Some(*val),
-            ExprKind::CharLit(c) => Some(*c as u8 as i8 as i128),
+            ExprKind::CharLit(c) => Some(*c as i128),
             // Float literals are constant, return truncated value
             ExprKind::FloatLit(val) => Some(val.to_f64() as i128),
